+++ depthai_ros_driver/include/depthai_ros_driver/utils.hpp
@@ -9,7 +9,6 @@
 
 namespace dai {
 enum class CameraBoardSocket;
-struct CameraFeatures;
 }  // namespace dai
 
 namespace depthai_ros_driver {
@@ -29,6 +28,6 @@ T getValFromMap(const std::string& name, const std::unordered_map<std::string, T
     }
 }
 std::string getUpperCaseStr(const std::string& string);
-std::string getSocketName(dai::CameraBoardSocket socket, std::vector<dai::CameraFeatures> camFeatures);
+std::string getSocketName(dai::CameraBoardSocket socket);
 }  // namespace utils
 }  // namespace depthai_ros_driver